backend/pyproject.toml
[tool.poetry]
name = "ventserver"
version = "0.0.0"
description = "Pufferfish ventilator backend server."
authors = []
packages = [
  { include = "ventserver"}
]

[tool.poetry.scripts]
serve = 'ventserver.server:serve'

[tool.poetry.dependencies]
# Language
python = "^3.7"
attrs = "^19.3.0"
# Protocols
cobs = { version = "^1.1.3" }
betterproto = { version = "^1.2.3" }
# I/O: Flow Control
trio = { version = "^0.14.0" }
trio-typing = { version = "^0.4.0" }
# I/O: Devices
pyserial = { version = "^3.4" }
trio-websocket = { version = "^0.8.0" }
# Deploy
click = "^7.1.2"

[tool.poetry.dev-dependencies]
# Code Quality
pylint = "^2.4.4"
mypy = "^0.770"
# Testing
tox = "^3.14.6"
coverage = "^5.0.4"
pytest = "^5.4.1"
pytest-sugar = "^0.9.3"
pytest-cov = "^2.8.1"
hypothesis = "^5.14.0"
pytest-trio = "^0.5.2"
# Unit Test Support
crcmod = "^1.7"
# Profiling
yappi = "^1.2.5"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
